build: reset password, orders, cart, notify

+++ database/migrations/2023_03_01_023403_create_order_details_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained();
             $table->foreignId('order_id')->constrained();
-            $table->integer('price')->nullable();
+            $table->float('price')->nullable();
             $table->integer('quantity')->nullable();
             $table->timestamps();
         });
